Test: move ElementDOMSlot tests to their own file with modern APIs
The DOMSlot implementation is no longer tightly coupled to
LexicalElementNode, so relocate the ElementDOMSlot describe blocks into a
dedicated LexicalDOMSlot.test.tsx. While moving, build editors with
buildEditorFromExtensions/defineExtension, mount via setRootElement, and
declare the test node classes with $config instead of static getType /
clone / importJSON / exportJSON boilerplate.
